Don't assume groups[concourseci_web_group] exist
Current code fails if there are no hosts in the concourseci_web_group (when e.g. applying this playbook only to worker hosts). This holds true even if specifying a concourse_web_options/CONCOURSE_TSA_HOST. This change gives a default list with one element as the second argument to `.get`